Add export command to output oci tarball
- Ensure transfer containerd plugin is configured with nix snapshotter
- Preload example images on VM boot
- Update README and local scripts with latest image references
- Kubernetes still have issues retrieving locally imported images
