+++ modules/nixos/vm.nix
@@ -1,31 +1,31 @@
-{ lib, config, pkgs, modulesPath, ... }:
-{
+{ lib, config, pkgs, modulesPath, examples, ... }:
+let
+  preloadContainerdImages =
+    builtins.map
+      (image: image.copyToOCIArchive {})
+      (lib.attrValues examples);
+
+in {
   imports = [
     # Import qemu-vm directly to avoid using vmVariant since this config
     # is only intended to be used as a VM. Using vmVariant will emit assertion
     # errors regarding `fileSystems."/"` and `boot.loader.grub.device`.
     (modulesPath + "/virtualisation/qemu-vm.nix")
-    ./kubernetes-startup.nix
+    # ./kubernetes.nix
+    ./k3s.nix
   ];
 
   # Enable rootful & rootless nix-snapshotter. This also starts rootful &
   # rootless containerd respectively.
   services.nix-snapshotter = {
     enable = true;
-    rootless.enable = true;
     setContainerdSnapshotter = true;
+    inherit preloadContainerdImages;
   };
 
-  # Provision single node kubernetes listening on localhost.
-  services.kubernetes = {
-    roles = ["master" "node"];
-    masterAddress = "localhost";
-  };
-
-  # Allow non-root "admin" user to just use `kubectl`.
-  services.certmgr.specs.clusterAdmin.private_key.owner = "admin";
-  environment.sessionVariables = {
-    KUBECONFIG = "/etc/${config.services.kubernetes.pki.etcClusterAdminKubeconfig}";
+  services.nix-snapshotter.rootless = {
+    enable = true;
+    inherit preloadContainerdImages;
   };
 
   # Provide an example kubernetes config for redis using a nix-snapshotter
@@ -46,16 +46,20 @@
   ];
 
   users.users = {
-    admin = {
+    root = {
+      initialHashedPassword = null;
+      password = "root";
+    };
+    rootless = {
       isNormalUser = true;
       extraGroups = [ "wheel" ];
-      password = "admin";
-      group = "admin";
+      password = "rootless";
+      group = "rootless";
     };
   };
 
   virtualisation = {
-    memorySize = 2048;
+    memorySize = 4096;
     cores = 4;
     graphics = false;
     diskImage = null;
