--- script/k8s/hello.yml
@@ -1,8 +1,0 @@
-apiVersion: v1
-kind: Pod
-metadata:
-  name: hello
-spec:
-  containers:
-  - name: hello
-    image: ghcr.io/pdtpartners/hello
