+++ script/k8s/redis.yaml
@@ -7,7 +7,9 @@ metadata:
 spec:
   containers:
   - name: redis
-    image: hinshun/redis:nix
+    image: docker.io/hinshun/redis:nix
+    # image: ghcr.io/pdtpartners/redis
+    # imagePullPolicy: Never
     args: [ "--protected-mode", "no" ]
     ports:
       - containerPort: 6379
