+++ script/rootless/nix-snapshotter.sh
@@ -1,3 +1,3 @@
 REPO_DIR=$(git rev-parse --show-toplevel)
 
-./build/bin/nix-snapshotter --config ./build/nix-snapshotter/nix-snapshotter.toml
+./build/bin/nix-snapshotter --log-level debug --config ./build/nix-snapshotter/nix-snapshotter.toml
